Small change to HTML and main
Changed the HTML to display the users name if they are logged in.
In the main some of the render template did not take in the user session and was displaying links as if the user was not logged in. Now the links should stay the same instead of switching back to login and registration when users is already logged in.